+++ app/models/user.rb
@@ -44,6 +44,8 @@ class User < ApplicationRecord
   has_many :daily_contracts, through: :user_daily_contracts
   has_many :faction_contributions, dependent: :destroy
   has_many :system_quest_assignments, dependent: :destroy
+  has_many :user_seasons, dependent: :destroy
+  has_many :seasons, through: :user_seasons
 
   belongs_to :active_title, class_name: "ShopItem", optional: true
   belongs_to :active_avatar_item, class_name: "ShopItem", optional: true